제발ㅠㅠ

+++ routes/postPage.js
@@ -682,17 +682,19 @@ router.delete("/delete", async (req, res) => {
     const exist = verifyUser.dataValues.postImageUrl; // 현재 URL에 전달된 id값을 받아서 db찾음
     const url = exist.split("/"); // exist 저장된 fileUrl을 가져옴
     const delFileName = url[url.length - 1];
-    s3.deleteObject(
-      {
-        Bucket: "a-fo-bucket2",
-        Key: delFileName,
-      },
-      (err, data) => {
-        if (err) {
-          throw err;
+    if (delFileName !== "A-fo_default.jpg") {
+      s3.deleteObject(
+        {
+          Bucket: "a-fo-bucket2",
+          Key: delFileName,
+        },
+        (err, data) => {
+          if (err) {
+            throw err;
+          }
         }
-      }
-    );
+      );
+    }
 
     // const postImg = verifyUser.dataValues.postImageUrl;
     // fs.unlink(__dirname + `/../uploads${postImg}`, (err) => {
